+++ better_exceptions/core/encoding.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 import codecs
 import locale
 import sys
