+++ setup.py
@@ -1,13 +1,14 @@
 import re
 from distutils.core import setup
 
-with open('better_exceptions/__init__.py', 'r') as file:
+with open('better_exceptions/core/__init__.py', 'r') as file:
     version = re.search(r'^__version__\s*=\s*[\'"]([^\'"]*)[\'"]',
                         file.read(), re.MULTILINE).group(1)
 
 setup(
     name = 'better_exceptions',
-    packages = ['better_exceptions'],
+    packages = ['better_exceptions', 'better_exceptions.core', 'better_exceptions_core'],
+    package_dir = {'better_exceptions_core': 'better_exceptions/core'},
     version = version,
     description = 'Pretty and helpful exceptions, automatically',
     author = 'Josh Junon',
